backend/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
pymongo==3.12.3
djongo==1.3.6
dnspython==2.1.0
python-dotenv==0.20.0